up to 8 chunks

+++ IRHermesInt.h
@@ -62,7 +62,7 @@ typedef
 		uint8_t       recvpin;         // Pin connected to IR data from detector
 		uint8_t       blinkpin;
 		uint8_t       blinkflag;       // true -> enable blinking of pin on IR processing
-		uint8_t       rawlen;          // counter of entries in rawbuf
+		uint16_t      rawlen;          // counter of entries in rawbuf
 		unsigned int  timer;           // State timer, counts 50uS ticks.
 		unsigned int  rawbuf[RAWBUF];  // raw data
 		uint8_t       overflow;        // Raw buffer overflow occurred
